split slider logic into separate file
