+++ src/elements/toniq-slider/toniq-slider.test.ts
@@ -5,7 +5,8 @@ import {createFixtureTest} from '../../element-testing/fixture-test';
 import {createElementRegistrationTest} from '../../element-testing/test-element-creation';
 import {createFocusTests} from '../../element-testing/test-focus';
 import {getByTestId} from '../../element-testing/test-id-testing';
-import {sliderTestIds, ToniqSlider} from './toniq-slider.element';
+import {sliderTestIds} from './slider-logic';
+import {ToniqSlider} from './toniq-slider.element';
 
 describe(ToniqSlider.tagName, () => {
     createElementRegistrationTest(ToniqSlider);
